Post and Put Validations Error Handler

+++ src/idea/idea.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { IdeaEntity } from './idea.entity';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -21,18 +21,33 @@ async create(data:IdeaDTO){
 }
 
 async read(id:string){
-    return await this.ideaRepository.find({where:{id}})
+    const idea = await this.ideaRepository.find({where:{id}});
+    if(!idea){
+        throw new HttpException('Not Found', HttpStatus.NOT_FOUND)
+    }
+    return idea;
 }
 
 async update(id:string,data:Partial<IdeaDTO>){
+    let idea = await this.ideaRepository.findOne({where:{id}});
+    if(!idea){
+        throw new HttpException("Not Found", HttpStatus.NOT_FOUND)
+    }
     await this.ideaRepository.update({id},data)
-    return await this.ideaRepository.findOne({id});
+    idea = await this.ideaRepository.findOne({where:{id}});
+    return idea;
 }
 
 
 async destroy(id:string){
+    const idea = await this.ideaRepository.findOne({where:{id}});
+
+    if(!idea){
+        throw new HttpException("Not Found", HttpStatus.NOT_FOUND)
+    }
+    
     await this.ideaRepository.delete({id});
-    return {delete:true}
+    return idea;
 }
